requirements.txt
django
django-rest-framework
dotenv
psycopg2
django-cors-headers
djangorestframework-simplejwt
djoser
